requirements.txt
Flask==2.3.3
tensorflow==2.15.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
Werkzeug==2.3.7